FW_Update_app/requirements.txt
pyserial
python-multipart
fastapi
uvicorn
